Fixed rank calculation for SemVers
SemVer and related functions were also moved to a new file.

The new rank calculation shows a flaw with the current requirements. Most people use the ^ prefix, which allows any minor or bug version greated than what is requested to be used. However, this doesn't require the newest version to be used, nor is it necessarily reflected in the deployed version. Currently, if ^ is used it lowers the severity by a rank.