ak-peter/traffic-light (#2)
* Added a tree view page for displaying dependencies.

This relies on a json file, cachedData.json, currently generated by repocrawler.

* Added the inverse tree view mode, where a repo lists what other repos depend on it.

* Added the combined view, which shows both dependencies and users.

* Slight refactor to where JSON data is loaded.

This is now processed in the function "JSObjectFromJSON", so any transformations can be added here.
It may be best to move this to a separate file.

* converted the app to typescript

* Added links and improved the data interface

The changes to the interface were needed as the JSON file was improved, but this also fixed the issue brough up in the meeting (that values were accessed by index rather than being converted to a proper JavaScript/TypeScript object).

* merging typescript branch

* Created Data-Table with Tabs template.

- Installed MUI

* Added the html links to the page

* Fixed CSS error, removed default table header

* 1 major behind is now red, commit before merge

* Merged ak-leanne-alex

* Renamed file for merging

* mend

* Finished merging and began abstracting components so styling can be worked on independently of data processing

* Customized styling of table to match Mockup

* Modified border widths

* minor changes to font

* Refactoring dataProcessing file

* fixed 'jsobObject0' not iterable error

* Brought the new styling in parity with the proof of concept

Added a subrow creater
Added rank calcualtions
Added row and subrow colouring based on rank
Added inverse dependencies to DependencyData
Added a "Users" tab

Moved red/green/orange colour information to css
Changed DependecyMap to use SemVer instead of a string for version

* Fixed bug with rank calculation

where the version of the parent was used instead of the newest version of the current repository.

* Re-done the black borders with @reteppeter

Co-Authored-By: reteppeter <22411750+reteppeter@users.noreply.github.com>

* changed the light status to a circular icon

Need to fix the CSS

* Fixed minor bug, rank was of incorrect type

* Replaced table row colour with traffic light for internal/external

* Made changes according to feedback in PR

* Fixed rank calculation for SemVers

SemVer and related functions were also moved to a new file.

The new rank calculation shows a flaw with the current requirements. Most people use the ^ prefix, which allows any minor or bug version greated than what is requested to be used. However, this doesn't require the newest version to be used, nor is it necessarily reflected in the deployed version. Currently, if ^ is used it lowers the severity by a rank.

* Updated the page to support the new cachedData format

Moved page creation to a separte file
Added pages for python and ruby (the index defaults to using typescript/javascript)
Data processing no longer requires the json object array to be split up when calling it

* test commit

* Test commit (testing commit without pre-commit hooks)

* Added support for ~> (from RubyGems) to the SemVer parser

* test commit

* fixing lint errors

* Removed old rendering code and fixed type issues

* Fixed typing and style loading issues, worked on colour configuration

Also added a new test page, /dark, which has a button to enable dark mode. The intent is to determine how we will switch styles.

* Running pre-commit hooks

Co-authored-by: akshithP <akshithp13@gmail.com>
Co-authored-by: reteppeter <22411750+reteppeter@users.noreply.github.com>
Co-authored-by: MT <59728838+mt-empty@users.noreply.github.com>